fix(auth): consolidate LanguageSelect, useLocation, bodyStyle, i18n Support
- refactor: merge LanguageSelect into LanguageToggle variant="select"
  (deletes 81-line duplicate, adds clickIdRef race guard to select)
- fix: replace window.location.pathname with useLocation() in OlaOs
  (subscribes to React Router for correct client-side re-renders)
- fix: replace deprecated bodyStyle with styles on Auth Card (AntD 5)
- fix: i18n "Support" link via translate('support') in AuthLayout footer
- chore: add support key to en.js/zh.js translation dictionaries
